native_table_detector/requirements.txt
paddleocr
pymupdf
opencv-python
numpy
pillow
gradio
opendataloader-pdf

